Add enumerate support
Both itemize and enumerate now use help of a state stack. That way
there are no global variables needed. For itemize it is saving
special bullet points and for enumerate it is saving the number
of current item. Enumerate uses almost the same principle as
itemize, but doesn't have it's own block yet.