UI file refactor and adding text editor
The variable naming in the ui.py file were a bit generic and
unreadable, so I have done a small naming refactor.

Additionally, there is a new feature added - Text Editor. Now,
the user is able to open text editor for the latex text and click
on the save & return button to update the changes in the addon.