Adding template for BLOCK, better itemize
I have decided that pure LL(1) would create too much duplicate
code, so this is the first step to LL(1) with semantic predicates.
The block template should be working for all of the block types I
have so far.

Additionally, I have remade the itemize and enumerate structure.
Now the LL table includes both of them under <ITEMIZE>, there is
context when working with them and less duplicate code.